# docs/orders_web_service_spring-openapi.yaml
openapi: "3.0.3"
info:
  title: "orders_web_service_spring API"
  description: "orders_web_service_spring API"
  version: "1.0.0"
servers:
  - url: "https://orders_web_service_spring"
paths:
  /orders/{id}/complete:
    put:
      summary: "PUT orders/{id}/complete"
      operationId: "complete"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /orders:
    post:
      summary: "POST orders"
      operationId: "newOrder"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
    get:
      summary: "GET orders"
      operationId: "all"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    delete:
      summary: "DELETE orders/{id}/cancel"
      operationId: "cancel"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /orders{id}:
    get:
      summary: "GET orders{id}"
      operationId: "one"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
components:
  schemas:
    Order:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        description:
          type: "string"
        status:
          type: "string"
          enum:
            - "IN_PROGRESS"
            - "COMPLETED"
            - "CANCELLED"